+++ FOX/functions/multi_mol.py
@@ -5,459 +5,16 @@ __all__ = ['MultiMolecule']
 from itertools import chain
 
 import numpy as np
+import pandas as pd
 from scipy.spatial.distance import cdist
 
 from scm.plams.mol.atom import Atom
 from scm.plams.mol.bond import Bond
 from scm.plams.mol.molecule import Molecule
-from scm.plams.core.settings import Settings
+from scm.plams.tools.periodic_table import PeriodicTable
 
-from .read_xyz import read_multi_xyz
 from .rdf import (get_rdf, get_rdf_lowmem, get_rdf_df)
-
-
-class _MultiMolecule:
-    """ A class for handling the magic methods of *MultiMolecule*. """
-    def __init__(self, coords=None, atoms=None, bonds=None, properties=None, filename=None,
-                 inputformat=None):
-        # Sanitize arguments
-        coords, atoms, bonds, properties = self._sanitize_init(coords, atoms, bonds, properties)
-
-        # Set attributes
-        self.coords = coords
-        self.atoms = atoms
-        self.bonds = bonds
-        self.properties = properties
-
-        # If **filename** is not *None*, grab coordinates from **filename**
-        if filename is not None:
-            format_dict = {'xyz': read_multi_xyz}
-
-            # Try to infere the inputformat from **filename** if **inputformat** is *None*
-            if inputformat is None:
-                try:
-                    inputformat = filename.rsplit('.', 1)[-1]
-                except AttributeError:
-                    error = 'MultiMolecule: Inputformat is None and no extension was found in '
-                    error += '**filename**'
-                    raise TypeError(error)
-
-            # Read **filename** if its inputformat is supported
-            if inputformat.lower().rsplit('.', 1)[-1] in format_dict:
-                self._read(filename, inputformat, format_dict)
-            else:
-                error = 'MultiMolecule: No functions are available for handling ' + str(inputformat)
-                error += ' files, currently supported file types are: '
-                error += str(list(format_dict.keys()))
-                raise KeyError(error)
-
-    def _read(self, filename, inputformat, format_dict):
-        """ A function for reading coordinates from external files. """
-        coords, atoms = format_dict[inputformat](filename)
-        self.coords = coords
-        self.atoms = atoms
-
-    def _sanitize_init(self, coords=None, atoms=None, bonds=None, properties=None):
-        """ A function for sanitizing the arguments of __init__(). """
-        # If coords is another MultiMolecule object, create a shallow copy and return
-        if isinstance(coords, _MultiMolecule):
-            properties = coords.properties
-            bonds = coords.bonds
-            atoms = coords.atoms
-            coords = coords.coords
-            return coords, atoms, bonds, properties
-
-        # Sanitize coords
-        assert coords is None or isinstance(coords, np.ndarray)
-        if coords is not None:
-            if coords.ndim == 2:
-                coords = coords[None, :, :]
-            elif not coords.ndim == 3:
-                raise TypeError('MultiMolecule: **coords** should be None or a 2D/3D array.')
-
-        # Sanitize atoms & properties
-        assert atoms is None or isinstance(atoms, dict)
-        assert properties is None or isinstance(properties, dict)
-        if isinstance(atoms, dict):
-            atoms = Settings(atoms)
-        properties = properties or Settings()
-
-        # Sanitize bonds
-        assert bonds is None or isinstance(bonds, np.ndarray)
-        if bonds is not None:
-            assert bonds.ndim == 2 and bonds.shape[1] in (2, 3)
-            assert bonds.dtype.type in (np.int8, np.int16, np.int32, np.int64)
-            if bonds.shape[1] == 2:
-                bonds = np.hstack((bonds, np.zeros(len(bonds), dtype=int)[:, None]))
-
-        return coords, atoms, bonds, properties
-
-    def _sanitize_other(self, other):
-        """ A function for sanitizing the argument **other**, an argument which appears in many
-        magic methods. """
-        # Validate object type and shape
-        try:
-            # Check for arrays
-            assert other.shape[-1] == 3
-
-            # Broadcast (if necessary) and return
-            if other.ndim == 2:
-                other = other[None, :, :]
-            elif other.ndim == 1:
-                other = other[None, None, :]
-            return other
-        except AttributeError:
-            # Assume other is a float or int
-            return other
-
-    def _set_shape(self, value): self.coords.shape = value
-    def _get_shape(self): return self.coords.shape
-    shape = property(_get_shape, _set_shape)
-
-    def _set_dtype(self, value): self.coords.dtype = value
-    def _get_dtype(self): return self.coords.dtype
-    dtype = property(_get_dtype, _set_dtype)
-
-    def _get_dtype(self): return self.coords.flags
-    flags = property(_get_dtype, _set_dtype)
-
-    def _get_dtype(self): return self.coords.ndim
-    ndim = property(_get_dtype)
-
-    def _get_dtype(self): return self.coords.nbytes
-    nbytes = property(_get_dtype)
-
-    """ ############################  Comparison magic methods  ############################### """
-
-    def __eq__(self, other):
-        other = self._sanitize_other(other)
-        return self.coords == other
-
-    def __ne__(self, other):
-        other = self._sanitize_other(other)
-        return self.coords != other
-
-    def __lt__(self, other):
-        other = self._sanitize_other(other)
-        return self.coords < other
-
-    def __gt__(self, other):
-        other = self._sanitize_other(other)
-        return self.coords > other
-
-    def __le__(self, other):
-        other = self._sanitize_other(other)
-        return self.coords <= other
-
-    def __ge__(self, other):
-        other = self._sanitize_other(other)
-        return self.coords >= other
-
-    """ ########################### Unary operators and functions  ############################ """
-
-    def __min__(self, axis=None):
-        return self.coords.nanmin(axis)
-
-    def __max__(self, axis=None):
-        return self.coords.nanmax(axis)
-
-    def __pos__(self):
-        return self.coords
-
-    def __neg__(self):
-        return -1 * self.coords
-
-    def __abs__(self):
-        return np.abs(self.coords)
-
-    def __round__(self, decimals=0):
-        return np.round(self.coords, decimals)
-
-    def __floor__(self):
-        return np.floor(self.coords)
-
-    def __ceil__(self):
-        return np.ceil(self.coords)
-
-    def __trunc__(self):
-        return np.trunc(self.coords)
-
-    """ ##########################  Normal arithmetic operators  ############################## """
-
-    def __add__(self, other):
-        other = self._sanitize_other(other)
-        ret = self.copy()
-        ret.coords = self.coords + other
-        return ret
-
-    def __sub__(self, other):
-        other = self._sanitize_other(other)
-        ret = self.copy()
-        ret.coords = self.coords - other
-        return ret
-
-    def __mul__(self, other):
-        other = self._sanitize_other(other)
-        ret = self.copy()
-        ret.coords = self.coords * other
-        return ret
-
-    def __matmul__(self, other):
-        other = self._sanitize_other(other)
-        ret = self.copy()
-        ret.coords = self.coords @ np.swapaxes(other, -2, -1)
-        return ret
-
-    def __floordiv__(self, other):
-        other = self._sanitize_other(other)
-        ret = self.copy()
-        ret.coords = self.coords // other
-        return ret
-
-    def __div__(self, other):
-        other = self._sanitize_other(other)
-        ret = self.copy()
-        ret.coords = self.coords / other
-        return ret
-
-    def __mod__(self, other):
-        other = self._sanitize_other(other)
-        ret = self.copy()
-        ret.coords = self.coords % other
-        return ret
-
-    def __divmod__(self, other):
-        other = self._sanitize_other(other)
-        ret = self.copy()
-        ret.coords = np.divmod(self.coords, other)
-        return ret
-
-    def __pow__(self, other):
-        other = self._sanitize_other(other)
-        ret = self.copy()
-        ret.coords = self.coords**other
-        return ret
-
-    """ ##########################  Reflected arithmetic operators  ########################### """
-
-    def __rsub__(self, other):
-        other = self._sanitize_other(other)
-        ret = self.copy()
-        ret.coords = other - self.coords
-        return ret
-
-    def __rfloordiv__(self, other):
-        other = self._sanitize_other(other)
-        ret = self.copy()
-        ret.coords = other // self.coords
-        return ret
-
-    def __rdiv__(self, other):
-        other = self._sanitize_other(other)
-        ret = self.copy()
-        ret.coords = other / self.coords
-        return ret
-
-    def __rmod__(self, other):
-        other = self._sanitize_other(other)
-        ret = self.copy()
-        ret.coords = other % self.coords
-        return ret
-
-    def __rdivmod__(self, other):
-        other = self._sanitize_other(other)
-        ret = self.copy()
-        ret.coords = np.divmod(other, self.coords)
-        return ret
-
-    def __rpow__(self, other):
-        other = self._sanitize_other(other)
-        ret = self.copy()
-        ret.coords = other**self.coords
-        return ret
-
-    """ ##############################  Augmented assignment  ################################# """
-
-    def __iadd__(self, other):
-        other = self._sanitize_other(other)
-        self.coords += other
-        return self
-
-    def __isub__(self, other):
-        other = self._sanitize_other(other)
-        self.coords -= other
-        return self
-
-    def __imul__(self, other):
-        other = self._sanitize_other(other)
-        self.coords *= other
-        return self
-
-    def __imatmul__(self, other):
-        other = self._sanitize_other(other)
-        self.coords = self.coords @ np.swapaxes(other, -2, -1)
-        return self
-
-    def __ifloordiv__(self, other):
-        other = self._sanitize_other(other)
-        self.coords //= other
-        return self
-
-    def __idiv__(self, other):
-        other = self._sanitize_other(other)
-        self.coords /= other
-        return self
-
-    def __imod__(self, other):
-        other = self._sanitize_other(other)
-        self.coords %= other
-        return self
-
-    def __ipow__(self, other):
-        other = self._sanitize_other(other)
-        self.coords **= other
-        return self
-
-    """ ##########################  Type conversion magic methods  ############################ """
-
-    def __str__(self):
-        # Convert atomic coordinates
-        ret = 'Atomic coordinates:\n'
-        if self.coords is not None:
-            ret += str(self.coords) + '\n'
-        else:
-            ret += str(None) + '\n'
-        ret += '\n'
-
-        # Convert atomic symbols
-        ret += 'Atomic symbols & indices:\n'
-        if self.atoms is not None:
-            for at in self.atoms:
-                ret += str(at) + ':\t['
-                if len(self.atoms[at]) <= 11:
-                    for i in self.atoms[at][0:-2]:
-                        ret += '{: <5.5}'.format(str(i))
-                else:
-                    for i in self.atoms[at][0:5]:
-                        ret += '{: <5.5}'.format(str(i))
-                    ret += '{: <5.5}'.format('...')
-                    for i in self.atoms[at][-6:-2]:
-                        ret += '{: <5.5}'.format(str(i))
-                ret += '{: <4.4}'.format(str(self.atoms[at][-1])) + ']\n'
-        else:
-            ret += str(None) + '\n'
-        ret += '\n'
-
-        # Convert bonds
-        ret += 'Bond indices and orders:\n'
-        if self.bonds is not None:
-            ret += 'Atom1 Atom2 Bond order'
-            for at1, at2, order in self.bonds:
-                ret += '[' + '{: <5.5}'.format(str(at1)) + '{: <5.5}'.format(str(at2)) + '] '
-                ret += str(order) + '\n'
-        else:
-            ret += str(None) + '\n'
-        ret += '\n'
-
-        # Convert properties
-        ret += 'Properties:\n'
-        if self.properties:
-            ret += str(self.properties) + '\n'
-        else:
-            ret += str(None) + '\n'
-
-        return ret
-
-    """ ################################  Custom Sequences  ################################### """
-
-    def __len__(self):
-        return len(self.coords)
-
-    def __getitem__(self, key):
-        if not isinstance(key, str):
-            return self.coords[key]
-        return self.atoms[key]
-
-    def __setitem__(self, key, value):
-        self.coords[key] = value
-
-    def __delitem__(self, key):
-        raise ValueError('cannot delete array elements')
-
-    def __iter__(self):
-        return iter(self.coords)
-
-    def __reversed__(self, axis=0):
-        ret = self.copy()
-        ret.coords = np.flip(self.coords, axis=axis)
-        return ret
-
-    def __contains__(self, item):
-        return item in self.coords
-
-    """ ###################################  Copy  ############################################ """
-
-    def __copy__(self, subset=None):
-        """ Magic method, create and return a new *MultiMolecule* and fill its attributes with
-        views of their respective counterparts in **self**.
-
-        :parameter subset: Copy a subset of attributes from **self**; if *None*, copy all
-            attributes. Accepts one or more of the following values as strings: *properties*,
-            *atoms* and/or *bonds*.
-        :type subset: |None|_, |str|_ or |tuple|_ [|str|_]
-        """
-        subset = subset or ('atoms', 'bonds', 'properties')
-        ret = _MultiMolecule()
-
-        # Copy atoms
-        if 'atoms' in subset:
-            ret.coords = self.coords
-            ret.atoms = self.atoms
-
-        # Copy bonds
-        if 'bonds' in subset:
-            ret.bonds = self.bonds
-
-        # Copy properties
-        if 'properties' in subset:
-            ret.properties = self.properties
-
-        return ret
-
-    def __deepcopy__(self, subset=None):
-        """ Magic method, create and return a deep copy of **self**.
-
-        :parameter subset: Deepcopy a subset of attributes from **self**; if *None*, copy all
-            attributes. Accepts one or more of the following values as string:
-            *properties*, *atoms* and/or *bonds*.
-        :type subset: |None|_, |str|_ or |tuple|_ [|str|_]
-        """
-        subset = subset or ('atoms', 'bonds', 'properties')
-        ret = self.__copy__(subset=subset)
-
-        # Deep copy atoms
-        if 'atoms' in subset:
-            try:
-                ret.coords = self.coords.copy()
-                ret.atoms = self.atoms.copy()
-            except AttributeError:
-                pass
-
-        # Deep copy bonds
-        if 'bonds' in subset and self.bonds is not None:
-            try:
-                ret.bonds = self.bonds.copy()
-            except AttributeError:
-                pass
-
-        # Deep copy properties
-        if 'properties' in subset:
-            try:
-                ret.properties = self.properties.copy()
-            except AttributeError:
-                pass
-
-        return ret
+from .multi_mol_magic import _MultiMolecule
 
 
 class MultiMolecule(_MultiMolecule):
@@ -505,20 +62,20 @@ class MultiMolecule(_MultiMolecule):
             **self.coords**.
         :type atom_subset: |None|_ or |tuple|_ [|str|_]
         """
-        atom_subset = atom_subset or self.atoms.keys()
+        atom_subset = self.subset_to_idx(atom_subset)
         mol = self.as_Molecule(mol_subset=0, atom_subset=atom_subset)[0]
         mol.guess_bonds()
         self.from_Molecule(mol, subset='bonds')
 
-    def remove_random_coords(self, p=0.5, return_coords=False):
+    def remove_random_coords(self, p=0.5, inplace=False):
         """ Remove random molecules from **self.coords**.
         For each molecule, the probability of removal is equal to **p**.
         Performs an inplace modification of **self.coords** if **return_coords** is *False*.
 
         :parameter float p: The probability of remove a 2D slice from **self.coords**.
             Accepts values between 0.0 (0%) and 1.0 (100%).
-        :parameter bool return_coords: If *True*, return a view of the gathered 2D frames.
-            If *False*, perform an inplace modification **self.coords**, replacing it with a
+        :parameter bool inplace: If *False*, return a view of the gathered 2D frames.
+            If *True*, perform an inplace modification **self.coords**, replacing it with a
             view of the randomly gathered molecules.
         """
         if p <= 0.0 or p >= 1.0:
@@ -529,15 +86,143 @@ class MultiMolecule(_MultiMolecule):
         size = 1 or int(self.shape[0] / p)
         idx_range = np.arange(self.shape[0])
         idx = np.random.choice(idx_range, size)
-        if return_coords:
+        if not inplace:
             return self[idx]
         self.coords = self[idx]
 
     """ ################################## Root Mean Squared ################################## """
 
-    def rms(self, ref, mol_subset=None, atom_subset=None):
-        ret = np.linalg.norm(self - ref, axis=(1, 2)) / self.shape[1]
-        return np.linalg.norm(self - ref) / self.shape[0]
+    def init_rmsd(self, other, mol_subset=None, atom_subset=None):
+        """ """
+        # Figure out if the RMSD should be calculated as a single or multiple series
+        try:
+            if isinstance(atom_subset[0], str) and not isinstance(atom_subset, str):
+                loop = True
+            else:
+                loop = False
+        except TypeError:
+            loop = False
+
+        # Get the RMSD
+        if loop:
+            rmsd = np.array([self.get_rmsd(other, mol_subset, at) for at in atom_subset]).T
+        else:
+            rmsd = self.get_rmsd(other, mol_subset, atom_subset)
+
+        # Construct arguments for the dataframe
+        if loop:
+            columns = atom_subset
+        elif isinstance(atom_subset, str):
+            columns = [atom_subset]
+        else:
+            columns = ['RMSD']
+        index = mol_subset or np.arange(0, self.shape[0])
+
+        # Create, fill and return a dataframe with the RMSD
+        df = pd.DataFrame(data=rmsd, index=index, columns=columns)
+        df.columns.name = 'RMSD  /  Ångström'
+        df.index.name = 'XYZ frame'
+        return df
+
+    def init_rmsf(self, mol_subset=None, atom_subset=None):
+        """ """
+        # Figure out if the RMSD should be calculated as a single or multiple series
+        try:
+            if isinstance(atom_subset[0], str) and not isinstance(atom_subset, str):
+                loop = True
+            else:
+                loop = False
+        except TypeError:
+            loop = False
+
+        # Get the RMSF
+        if loop:
+            rmsf = np.array([self.get_rmsf(mol_subset, at) for at in atom_subset]).T
+        else:
+            rmsf = self.get_rmsf(mol_subset, atom_subset)
+
+        # Construct arguments for the dataframe
+        index = np.arange(0, self.shape[1])
+        if loop:
+            columns = atom_subset
+            data = np.empty((rmsf.shape[0], index.shape[0]))
+            data[:, :] = np.nan
+            for i, (at, j) in enumerate(zip(list(atom_subset), rmsf)):
+                idx = self.subset_to_idx(at)
+                data[i, idx] = j
+            data = data.T
+        elif isinstance(atom_subset, str):
+            columns = [atom_subset]
+            data = np.empty((index.shape[0]))
+            data[:] = np.nan
+            idx = self.subset_to_idx(atom_subset)
+            data[idx] = rmsf
+        else:
+            columns = ['RMSF']
+            data = rmsf
+
+        # Create, fill and return a dataframe with the RMSF
+        df = pd.DataFrame(data=data, index=index, columns=columns)
+        df.columns.name = 'RMSF  /  Ångström'
+        df.index.name = 'Atomic index'
+        return df
+
+    def get_rmsd(self, other, mol_subset=None, atom_subset=None):
+        """ Calculate the root mean square fluctuation (RMSF) between **self.coords** and
+        **other**. Returns a dataframe with the RMSD as a function of the XYZ frame numbers. """
+        # Prepare slices
+        i = mol_subset or slice(0, self.shape[0])
+        j = self.subset_to_idx(atom_subset) or slice(0, self.shape[1])
+
+        # Check if the shapes of self and other match
+        if self[i, j, :].shape == other[i, j, :].shape:
+            other = other[i, j, :]
+
+        # Calculate and return the RMSD per molecule in **self.coords**
+        dist = np.linalg.norm(self[i, j, :] - other, axis=2)
+        return np.sqrt(np.einsum('ij,ij->i', dist, dist) / dist.shape[1])
+
+    def get_rmsf(self, mol_subset=None, atom_subset=None):
+        """ Calculate the root mean square fluctuation (RMSF) of **self.coords**.
+        Returns a dataframe as a function of atomic indices. """
+        # Prepare slices
+        i = mol_subset or slice(0, self.shape[0])
+        j = self.subset_to_idx(atom_subset) or slice(0, self.shape[1])
+
+        # Calculate the RMSF per molecule in **self.coords**
+        mean_coords = np.average(self[i, j, :], axis=0)[None, :, :]
+        displacement = np.linalg.norm(self[i, j, :] - mean_coords, axis=2)
+        return np.linalg.norm(displacement, axis=0)
+
+    def reset_origin(self, mol_subset=None, atom_subset=None, inplace=True):
+        """ Set the origin to the center of mass of **self.coords**. Performs in inplace update
+        of **self.coords** if **inplace** is *True*. """
+        center_of_mass = self.get_center_of_mass(mol_subset, atom_subset)
+        if not inplace:
+            return self - center_of_mass[:, None, :]
+        self -= center_of_mass[:, None, :]
+
+    def get_center_of_mass(self, mol_subset=None, atom_subset=None):
+        """ Return the center of mass of *m* molecules as an *m*3* array. """
+        # Prepare a dictionary with atomic masses
+        mass_dict = {}
+        for at in self.atoms:
+            mass = PeriodicTable.get_mass(at)
+            for i in self.atoms[at]:
+                try:
+                    mass_dict[i].append(mass)
+                except KeyError:
+                    mass_dict[i] = [mass]
+
+        # Prepare slices
+        i = mol_subset or slice(0, self.shape[0])
+        j = self.subset_to_idx(atom_subset) or slice(0, self.shape[1])
+
+        # Get the center of mass
+        mass = np.array([mass_dict[i] for i in atom_subset])[None, :]
+        weighted_coords = self[i, j, :] * mass[:, None]
+        center = weighted_coords.sum(axis=2)
+        return center / mass.sum(axis=1)
 
     """ ###################  Radial and Angular Distribution Functions  ####################### """
 
@@ -610,8 +295,8 @@ class MultiMolecule(_MultiMolecule):
         """
         # Define array slices
         mol_subset = mol_subset or slice(0, self.shape[0])
-        i = mol_subset, self._get_idx(atom_subset[0]) or slice(0, self.shape[1])
-        j = mol_subset, self._get_idx(atom_subset[1]) or slice(0, self.shape[1])
+        i = mol_subset, self.subset_to_idx(atom_subset[0]) or slice(0, self.shape[1])
+        j = mol_subset, self.subset_to_idx(atom_subset[1]) or slice(0, self.shape[1])
 
         # Slice the XYZ array
         A = self[i]
@@ -632,7 +317,7 @@ class MultiMolecule(_MultiMolecule):
             **self.coords**. If *None*, create a distance matrix for all molecules in
             **self.coords**.
         :type mol_subset: |None|_ or |tuple|_ [|int|_]
-       :parameter atom_subset: Create a distance matrix from a subset of atoms per molecule in
+        :parameter atom_subset: Create a distance matrix from a subset of atoms per molecule in
             **self.coords**. Values have to be supplied for all 3 dimensions. Atomic indices
             (on or multiple), atomic symbols (one or multiple) and *None* can be freely mixed.
             If *None*, pick all atoms from **self.coords** for that partical dimension; if an
@@ -644,9 +329,9 @@ class MultiMolecule(_MultiMolecule):
         """
         # Define array slices
         mol_subset = mol_subset or slice(0, self.shape[0])
-        i = mol_subset, self._get_idx(atom_subset[0]) or slice(0, self.shape[1])
-        j = mol_subset, self._get_idx(atom_subset[1]) or slice(0, self.shape[1])
-        k = mol_subset, self._get_idx(atom_subset[2]) or slice(0, self.shape[1])
+        i = mol_subset, self.subset_to_idx(atom_subset[0]) or slice(0, self.shape[1])
+        j = mol_subset, self.subset_to_idx(atom_subset[1]) or slice(0, self.shape[1])
+        k = mol_subset, self.subset_to_idx(atom_subset[2]) or slice(0, self.shape[1])
 
         # Slice and broadcast the XYZ array
         A = self[i][:, None, :]
@@ -663,17 +348,17 @@ class MultiMolecule(_MultiMolecule):
         ret = np.arccos(np.einsum('ijkl,ijml->ijkl', unit_vec1, unit_vec2))
         return ret
 
-    def _get_idx(self, at):
+    def subset_to_idx(self, arg):
         """ Grab and return a list of indices from **self.atoms**.
         Return *at* if it is *None*, an *int* or iterable container consisting of *int*. """
-        if at is None:
-            return at
-        elif isinstance(at, int) or isinstance(at[0], int):
-            return at
-        elif isinstance(at, str):
-            return self.atoms[at]
-        elif isinstance(at[0], str):
-            return list(chain.from_iterable(self.atoms[i] for i in at))
+        if arg is None:
+            return None
+        elif isinstance(arg, int) or isinstance(arg[0], int):
+            return arg
+        elif isinstance(arg, str):
+            return self.atoms[arg]
+        elif isinstance(arg[0], str):
+            return list(chain.from_iterable(self.atoms[i] for i in arg))
         raise KeyError()
 
     """ #################################  Type conversion  ################################### """
@@ -691,15 +376,16 @@ class MultiMolecule(_MultiMolecule):
         :rtype: |list|_ [|plams.Molecule|_].
         """
         # Create a dictionary with atomic indices as keys and matching atomic symbols as values
-        atom_subset = atom_subset or self.atoms.values()
-        at_list = list(chain.from_iterable(atom_subset))
+        mol_subset = mol_subset or np.arange(1, self.shape[0])
+        atom_subset = self.subset_to_idx(atom_subset)
+        atom_subset = atom_subset or list(chain.from_iterable(self.atoms.values()))
 
         # Construct a template molecule and fill it with atoms
         assert self.coords is not None
         assert self.atoms is not None
         mol_template = Molecule()
         mol_template.properties = self.properties.copy()
-        for i, symbol in enumerate(at_list):
+        for i, symbol in enumerate(atom_subset):
             atom = Atom(symbol=symbol)
             mol_template.add_atom(atom)
 
@@ -714,8 +400,7 @@ class MultiMolecule(_MultiMolecule):
 
         # Create copies of the template molecule; update their cartesian coordinates
         ret = []
-        idx_range = mol_subset or np.arange(1, self.shape[0])
-        for i, xyz in zip(self, idx_range):
+        for i, xyz in zip(self, mol_subset):
             mol = mol_template.copy()
             mol.from_array(xyz)
             mol.properties.frame = i
